pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leadflow"
version = "0.1.0"
description = "CLI-based lead enrichment pipeline — Clay.com replacement"
requires-python = ">=3.12"
dependencies = [
    "supabase>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
leadflow = "leadflow.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
